// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/fasting-message-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface FastingMessageHttpResponse {
    bodyKey?: string;
    bodyParams?: { [key: string]: string } | null;
    id?: string;
    titleKey?: string;
    tone?: string;
}