// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/fasting-overview-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FastingInsightsHttpResponse } from './fasting-insights-http-response';
import { FastingSessionHttpResponse } from './fasting-session-http-response';
import { FastingSessionHttpResponsePagedHttpResponse } from './fasting-session-http-response-paged-http-response';
import { FastingStatsHttpResponse } from './fasting-stats-http-response';

export interface FastingOverviewHttpResponse {
    currentSession?: FastingSessionHttpResponse | null;
    history?: FastingSessionHttpResponsePagedHttpResponse;
    insights?: FastingInsightsHttpResponse;
    stats?: FastingStatsHttpResponse;
}