// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/fasting-session-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FastingCheckInHttpResponse } from './fasting-check-in-http-response';

export interface FastingSessionHttpResponse {
    addedDurationHours?: number;
    checkInAtUtc?: string | null;
    checkInNotes?: string | null;
    checkIns?: Array<FastingCheckInHttpResponse>;
    cyclicEatDayEatingWindowHours?: number | null;
    cyclicEatDayFastHours?: number | null;
    cyclicEatDays?: number | null;
    cyclicFastDays?: number | null;
    cyclicPhaseDayNumber?: number | null;
    cyclicPhaseDayTotal?: number | null;
    endedAtUtc?: string | null;
    energyLevel?: number | null;
    hungerLevel?: number | null;
    id?: string;
    initialPlannedDurationHours?: number;
    isCompleted?: boolean;
    moodLevel?: number | null;
    notes?: string | null;
    occurrenceKind?: string;
    planType?: string;
    plannedDurationHours?: number;
    protocol?: string;
    startedAtUtc?: string;
    status?: string;
    symptoms?: Array<string>;
}